Remove text renderer API from core and add renderer submodule
